LPS-62487 Rename bundle and package name
