+++ modules/apps/export-import/export-import-resources-importer/bnd.bnd
@@ -1,5 +1,5 @@
-Bundle-Name: Liferay Resources Importer
-Bundle-SymbolicName: com.liferay.resources.importer
+Bundle-Name: Liferay Export Import Resources Importer
+Bundle-SymbolicName: com.liferay.exportimport.resources.importer
 Bundle-Version: 1.0.0
 Liferay-Releng-Module-Group-Description:
 Liferay-Releng-Module-Group-Title: Data Management
