+++ modules/apps/export-import/export-import-resources-importer/src/main/java/com/liferay/exportimport/resources/importer/util/PluginPackageProperties.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.resources.importer.util;
+package com.liferay.exportimport.resources.importer.util;
 
 import com.liferay.portal.kernel.util.GetterUtil;
 import com.liferay.portal.kernel.util.PortalRunMode;
